% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get.R
\name{get_categories}
\alias{get_categories}
\title{get category metadata or data from The Setup}
\usage{
get_categories(category = NULL, ...)
}
\arguments{
\item{category}{the name(s) of the categories; this can be found by looking at the
"categories" field in the return from \code{\link{get_interviews}}. Alternately, in the absence
of any provided categories, the /meta/data associated with all categories
on The Setup will instead be returned, and this includes the names.}

\item{...}{further arguments to pass to httr's GET.}
}
\description{
retrieves either the metadata about the interviewee categories on
The Setup, or the data from interviewees in particular categories, depending on
the provided input.
}
\examples{
#Get ALL the metadata!
all_categories <- get_categories()

#Get the information for actors
actors <- get_categories("actor")
}
\seealso{
\code{\link{get_software}} and \code{\link{get_interviews}}, for similar functionality
for hardware and full interviews.
}

