% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get.R
\name{get_hardware}
\alias{get_hardware}
\title{get The Setup's hardware data}
\usage{
get_hardware(item = NULL, ...)
}
\arguments{
\item{item}{the name of the specific hardware; this can be found by looking at the
hardware data in the return from \code{\link{get_interviews}}. Alternately, in the absence
of any provided hardware names, the /meta/data associated with all hardware entries
on The Setup will instead be returned, and this includes the names.}

\item{...}{further arguments to pass to httr's GET.}
}
\description{
retrieves either the metadata about hardware entries on The Setup, or the
data to do with a particular piece of hardware, depending on the provided input.
}
\examples{
#Get ALL the metadata!
all_hardware <- get_hardware()

#How many mac users does the site have?
photoshop_data <- get_hardware("macbook-pro")
#These people are a bit mac-happy, but it's okay, I'm sure they give generously to charity
#and are kind to small children. We can probably forgive them.
}
\seealso{
\code{\link{get_software}} and \code{\link{get_interviews}}, for similar functionality
for hardware and full interviews.
}

