% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get.R
\name{get_interviews}
\alias{get_interviews}
\title{Get The Setup's interview data}
\usage{
get_interviews(slugs = NULL, ...)
}
\arguments{
\item{slugs}{the "slug" of the specific interview, or interviews. For example, the slug of
Hadley's entry, which is listed at
\href{http://hadley.wickham.usesthis.com/}{http://hadley.wickham.usesthis.com/}, is
"hadley.wickham". If this is left empty, \code{get_interviews} will instead retrieve
the metadata associated with all of The Setup's interviews (which, appropriately,
includes all the slugs!)}

\item{...}{further arguments to httr's GET.}
}
\description{
retrieve either the metadata associated with each interview on The Setup, or
the metadata and data associated with a specific interview or set of interviews (see below)
}
\examples{
#Include a slug, retrieving the data associated with a particular interview
hadley_data <- get_interviews(slugs = "hadley.wickham")

#Or, sod it, don't and just get the metadata
setup_metadata <- get_interviews()
}
\seealso{
\code{\link{get_hardware}} and \code{\link{get_software}}, for similar functionality
for specific classes of data and metadata.
}

