% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get.R
\name{get_software}
\alias{get_software}
\title{get The Setup's software data}
\usage{
get_software(item = NULL, ...)
}
\arguments{
\item{item}{the name of the specific software; this can be found by looking at the
software data in the return from \code{\link{get_interviews}}. Alternately, in the absence
of any provided software names, the /meta/data associated with all software entries
on The Setup will instead be returned, and this includes the names}

\item{...}{further arguments to pass to httr's GET.}
}
\description{
retrieves either the metadata about software entries on The Setup, or the
data to do with a particular piece of software, depending on the provided nput.
}
\examples{
#Get ALL the metadata!
all_software <- get_software()

#Find out who is using photoshop (after that we should find out who switched it out for
#GIMP, and try to work out what the heck is wrong with them.)
photoshop_data <- get_software("photoshop")
}
\seealso{
\code{\link{get_hardware}} and \code{\link{get_interviews}}, for similar functionality
for hardware and full interviews.
}

